refactor: move purchaseItem to its own file
