+++ src/game-logic/reducers/purchaseItem.js
@@ -1,0 +1,50 @@
+import {
+  getAdjustedItemValue,
+  inventorySpaceRemaining,
+  moneyTotal,
+} from '../../utils'
+import { PURCHASED_ITEM_PEER_NOTIFICATION } from '../../templates'
+
+import { addItemToInventory } from './addItemToInventory'
+
+import { prependPendingPeerMessage } from './index'
+
+/**
+ * @param {farmhand.state} state
+ * @param {farmhand.item} item
+ * @param {number} [howMany=1]
+ * @returns {farmhand.state}
+ */
+export const purchaseItem = (state, item, howMany = 1) => {
+  const { money, todaysPurchases, valueAdjustments } = state
+  const numberOfItemsToAdd = Math.min(howMany, inventorySpaceRemaining(state))
+
+  if (numberOfItemsToAdd === 0) {
+    return state
+  }
+
+  const value = getAdjustedItemValue(valueAdjustments, item.id)
+  const totalValue = value * numberOfItemsToAdd
+
+  if (totalValue > money) {
+    return state
+  }
+
+  state = prependPendingPeerMessage(
+    state,
+    PURCHASED_ITEM_PEER_NOTIFICATION`${howMany}${item}`
+  )
+
+  return addItemToInventory(
+    {
+      ...state,
+      money: moneyTotal(money, -totalValue),
+      todaysPurchases: {
+        ...todaysPurchases,
+        [item.id]: (todaysPurchases[item.id] || 0) + numberOfItemsToAdd,
+      },
+    },
+    item,
+    numberOfItemsToAdd
+  )
+}
